+++ migrator.py
@@ -1,7 +1,7 @@
-from Migrators.mitre_attack.mitre_attack_migrator import migrate_mitre
+from migrators.mitre_attack.mitre_attack_migrator import migrate_mitre
 from typedb.client import *
 from timeit import default_timer as timer
-from Schema.initialise import initialise_database
+from schema.initialise import initialise_database
 
 uri = "localhost:1729"
 batch_size = 50
@@ -13,4 +13,4 @@ initialise_database(client, "cti", False)
 migrate_mitre(uri, batch_size, num_threads)
 end = timer()
 time_in_sec = end - start
-print("Elapsed time: " + str(time_in_sec) + " seconds.")
+print(f"Elapsed time: {time_in_sec} seconds.")
