+++ migrators/mitre_attack/concept_mapper.py
@@ -1,15 +1,14 @@
 def entity_mapper(entity: str):
-
     mapper = {
-    	"attack-pattern": {"type": "attack-pattern", "custom-type": False, 'ignore': False},
+        "attack-pattern": {"type": "attack-pattern", "custom-type": False, 'ignore': False},
         "tool": {"type": "tool", "custom-type": False, 'ignore': False},
         "identity": {"type": "identity", "custom-type": False, 'ignore': False},
         "course-of-action": {"type": "course-of-action", "custom-type": False, 'ignore': False},
         "malware": {"type": "malware", "custom-type": False, 'ignore': False},
         "intrusion-set": {"type": "intrusion-set", "custom-type": False, 'ignore': False},
-        "marking_definition": {"type": "marking-definition", "custom-type": False, 'ignore': True}    
-
+        "marking_definition": {"type": "marking-definition", "custom-type": False, 'ignore': True}
     }
+
     mapping = mapper.get(entity, {})
 
     if mapping == {}:
@@ -21,27 +20,28 @@ def entity_mapper(entity: str):
 
 
 def relationship_mapper(relationship: str):
-
     mapper = {
-    	"uses": {"relation-name": "use", "active-role": "used-by", "passive-role": "used"},
-    	"mitigates": {"relation-name": "mitigation", "active-role": "mitigating", "passive-role": "mitigated"},
+        "uses": {"relation-name": "use", "active-role": "used-by", "passive-role": "used"},
+        "mitigates": {"relation-name": "mitigation", "active-role": "mitigating", "passive-role": "mitigated"},
         "delivers": {"relation-name": "delivery", "active-role": "delivering", "passive-role": "delivered"},
         "targets": {"relation-name": "target", "active-role": "targetting", "passive-role": "targetted"},
         "attributed-to": {"relation-name": "attribution", "active-role": "attributing", "passive-role": "attributed"},
         "indicates": {"relation-name": "indication", "active-role": "indicating", "passive-role": "indicated"},
         "uses": {"relation-name": "use", "active-role": "used-by", "passive-role": "used"},
         "derives": {"relation-name": "derivation", "active-role": "deriving", "passive-role": "derived-from"},
-        "duplicate-of": {"relation-name": "duplicate", "active-role": "duplicate-object", "passive-role": "duplicate-object"},
+        "duplicate-of": {"relation-name": "duplicate", "active-role": "duplicate-object",
+                         "passive-role": "duplicate-object"},
         "related-to": {"relation-name": "relatedness", "active-role": "related-to", "passive-role": "related-to"}
     }
 
     mapping = mapper.get(relationship, {})
     if mapping == {}:
-        mapping = {"relation-name": "stix-core-relationship", "active-role": "active-role", "passive-role": "passive-role", "stix-type": relationship}
+        mapping = {"relation-name": "stix-core-relationship", "active-role": "active-role",
+                   "passive-role": "passive-role", "stix-type": relationship}
     return mapping
 
 
-def attribute_map():
+def attribute_definitions():
     mapper = {
         "id": {"type": "stix-id", "value": "string"},
         "created": {"type": "created", "value": "string"},
@@ -57,25 +57,3 @@ def attribute_map():
         "external_id": {"type": "external-id", "value": "string"}
     }
     return mapper
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
